apps/podcast-fetch-episode/src/requirements.txt
ffmpeg-python==0.2.0
google-cloud-speech==1.3.1
google-cloud-storage==1.24.1
